+++ tests/schedulers/test_scheduler_score_sde_ve.py
@@ -1,3 +1,9 @@
+import unittest
+
+import torch
+
+from diffusers import ScoreSdeVeScheduler
+
 
 class ScoreSdeVeSchedulerTest(unittest.TestCase):
     # TODO adapt with class SchedulerCommonTest (scheduler needs Numpy Integration)
@@ -179,4 +185,3 @@ class ScoreSdeVeSchedulerTest(unittest.TestCase):
 
             self.assertEqual(output_0.shape, sample.shape)
             self.assertEqual(output_0.shape, output_1.shape)
-
