+++ tests/schedulers/test_scheduler_vq_diffusion.py
@@ -1,3 +1,10 @@
+import torch
+
+from diffusers import VQDiffusionScheduler
+
+from .test_schedulers import SchedulerCommonTest
+
+
 class VQDiffusionSchedulerTest(SchedulerCommonTest):
     scheduler_classes = (VQDiffusionScheduler,)
